+++ packages/amplify-cli/src/extensions/amplify-helpers/get-provider-plugins.ts
@@ -1,7 +1,10 @@
-import { $TSContext, stateManager } from 'amplify-cli-core';
+import { $TSAny, $TSContext, stateManager } from 'amplify-cli-core';
 import _ from 'lodash';
 
-export function getProviderPlugins(context: $TSContext): Record<string, string> {
+/**
+ * Get the provider plugins
+ */
+export const getProviderPlugins = (context: $TSContext): Record<string, string> => {
   const providers = {};
   context.runtime.plugins.forEach(plugin => {
     if (plugin.pluginType === 'provider') {
@@ -9,9 +12,12 @@ export function getProviderPlugins(context: $TSContext): Record<string, string> 
     }
   });
   return providers;
-}
+};
 
-export const getConfiguredProviders = (context: $TSContext) => {
+/**
+ * Get configured providers
+ */
+export const getConfiguredProviders = (context: $TSContext): Record<string, string> => {
   const configuredProviders = stateManager.getProjectConfig()?.providers;
   if (!Array.isArray(configuredProviders) || configuredProviders.length < 1) {
     throw new Error('No providers are configured for the project');
@@ -19,7 +25,10 @@ export const getConfiguredProviders = (context: $TSContext) => {
   return _.pick(getProviderPlugins(context), configuredProviders) as Record<string, string>;
 };
 
-export const executeProviderCommand = async (context: $TSContext, command: string, args: unknown[] = []) => {
+/**
+ * Execute the provider command
+ */
+export const executeProviderCommand = async (context: $TSContext, command: string, args: unknown[] = []): Promise<$TSAny> => {
   const providers = await Promise.all(Object.values(getConfiguredProviders(context)).map(providerPath => import(providerPath)));
   await Promise.all(
     providers.filter(provider => typeof provider?.[command] === 'function').map(provider => provider[command](context, ...args)),