+++ packages/amplify-provider-awscloudformation/src/push-resources.ts
@@ -36,6 +36,8 @@ import {
   readCFNTemplate,
   Template,
   ApiCategoryFacade,
+  amplifyErrorWithTroubleshootingLink,
+  amplifyFaultWithTroubleshootingLink,
 } from 'amplify-cli-core';
 import ora from 'ora';
 import { Fn } from 'cloudform-types';
@@ -63,6 +65,7 @@ import { isAmplifyAdminApp } from './utils/admin-helpers';
 import { fileLogger } from './utils/aws-logger';
 import { APIGW_AUTH_STACK_LOGICAL_ID, loadApiCliInputs } from './utils/consolidate-apigw-policies';
 import { createEnvLevelConstructs } from './utils/env-level-constructs';
+import { showBuildDirChangesMessage } from './utils/auto-updates';
 import { NETWORK_STACK_LOGICAL_ID } from './network/stack';
 import { preProcessCFNTemplate, writeCustomPoliciesToCFNTemplate } from './pre-push-cfn-processor/cfn-pre-processor';
 import { AUTH_TRIGGER_STACK, AUTH_TRIGGER_TEMPLATE } from './utils/upload-auth-trigger-template';
@@ -82,8 +85,6 @@ const logger = fileLogger('push-resources');
 // keep in sync with ServiceName in amplify-category-api, but probably it will not change
 const ApiServiceNameElasticContainer = 'ElasticContainer';
 
-const spinner = ora('Updating resources in the cloud. This may take a few minutes...');
-
 const optionalBuildDirectoryName = 'build';
 const cfnTemplateGlobPattern = '*template*.+(yaml|yml|json)';
 const parametersJson = 'parameters.json';
@@ -98,7 +99,7 @@ const deploymentInProgressErrorMessage = () => {
 };
 
 /**
- *
+ * Prepares and pushes resources to the cloud.
  */
 export const run = async (context: $TSContext, resourceDefinition: $TSObject, rebuild = false) => {
   const deploymentStateManager = await DeploymentStateManager.createDeploymentStateManager(context);
@@ -135,11 +136,11 @@ export const run = async (context: $TSContext, resourceDefinition: $TSObject, re
     await createEnvLevelConstructs(context);
 
     // removing dependent functions if @model{Table} is deleted
-    const apiResourceTobeUpdated = resourcesToBeUpdated.filter((resource: { service: string; }) => resource.service === 'AppSync');
-    if (apiResourceTobeUpdated.length) {
+    const apiResourceToBeUpdated = resourcesToBeUpdated.filter((resource: { service: string; }) => resource.service === 'AppSync');
+    if (apiResourceToBeUpdated.length) {
       const functionResourceToBeUpdated = await ensureValidFunctionModelDependencies(
         context,
-        apiResourceTobeUpdated,
+        apiResourceToBeUpdated,
         allResources as $TSObject[],
       );
       // filter updated function to replace with existing updated ones(in case of duplicates)
@@ -193,8 +194,9 @@ export const run = async (context: $TSContext, resourceDefinition: $TSObject, re
     await prePushLambdaLayerPrompt(context, resources);
     await prepareBuildableResources(context, resources);
     await buildOverridesEnabledResources(context, resources);
+    await showBuildDirChangesMessage();
 
-    // Removed api transformation to generate resources befoe starting deploy/
+    // Removed api transformation to generate resources before starting deploy/
 
     // If there is a deployment already in progress we have to fail the push operation as another
     // push in between could lead non-recoverable stacks and files.
@@ -223,7 +225,7 @@ export const run = async (context: $TSContext, resourceDefinition: $TSObject, re
         const gqlManager = await GraphQLResourceManager.createInstance(context, gqlResource, cloudformationMeta.StackId, rebuild);
         deploymentSteps = await gqlManager.run();
 
-        // If any models are being replaced, we prepend steps to the iterative deployment to remove references to the replaced table in functions that have a dependeny on the tables
+        // If any models are being replaced, we prepend steps to the iterative deployment to remove references to the replaced table in functions that have a dependency on the tables
         const modelsBeingReplaced = gqlManager.getTablesBeingReplaced().map(meta => meta.stackName); // stackName is the same as the model name
         deploymentSteps = await prependDeploymentStepsToDisconnectFunctionsFromReplacedModelTables(
           context,
@@ -274,6 +276,7 @@ export const run = async (context: $TSContext, resourceDefinition: $TSObject, re
       // if there are deploymentSteps, need to do an iterative update
       if (deploymentSteps.length > 0) {
         // create deployment manager
+        const spinner = ora('Updating resources in the cloud. This may take a few minutes...');
         const deploymentManager = await DeploymentManager.createInstance(context, cloudformationMeta.DeploymentBucketName, spinner, {
           userAgent: formUserAgentParam(context, generateUserAgentAction(resourcesToBeCreated, resourcesToBeUpdated)),
         });
@@ -304,7 +307,7 @@ export const run = async (context: $TSContext, resourceDefinition: $TSObject, re
         });
         await deploymentManager.deploy(deploymentStateManager);
 
-        // delete the intermidiate states as it is ephemeral
+        // delete the intermediate states as it is ephemeral
         if (stateFolder.local) {
           try {
             fs.removeSync(stateFolder.local);
@@ -332,7 +335,10 @@ export const run = async (context: $TSContext, resourceDefinition: $TSObject, re
           if (err?.name === 'ValidationError' && err?.message === 'No updates are to be performed.') {
             return;
           }
-          throw err;
+          throw amplifyFaultWithTroubleshootingLink('DeploymentFault', {
+            stack: err.stack,
+            message: err.message,
+          });
         }
       }
       context.usageData.stopCodePathTimer('pushDeployment');
@@ -433,13 +439,13 @@ export const run = async (context: $TSContext, resourceDefinition: $TSObject, re
 
     await downloadAPIModels(context, newAPIresources);
 
-    // remove emphemeral Lambda layer state
+    // remove ephemeral Lambda layer state
     if (resources.concat(resourcesToBeDeleted).filter((r: { service: string; }) => r.service === AmplifySupportedService.LAMBDA_LAYER).length > 0) {
       await postPushLambdaLayerCleanup(context, resources, projectDetails.localEnvInfo.envName);
       await context.amplify.updateamplifyMetaAfterPush(resources);
     }
 
-    // Store current cloud backend in S3 deployment bcuket
+    // Store current cloud backend in S3 deployment bucket
     await storeCurrentCloudBackend(context);
     await amplifyServiceManager.storeArtifactsForAmplifyService(context);
 
@@ -456,15 +462,15 @@ export const run = async (context: $TSContext, resourceDefinition: $TSObject, re
       await deploymentStateManager.failDeployment();
     }
     rollbackLambdaLayers(layerResources);
-
-    logger('run', [resourceDefinition])(error);
-
-    throw error;
+    throw amplifyFaultWithTroubleshootingLink('DeploymentFault', {
+      stack: error.stack,
+      message: error.message,
+    });
   }
 };
 
 /**
- *
+ * update the cloudformation stack for the API migration
  */
 export const updateStackForAPIMigration = async (context: $TSContext, category: string, resourceName: string, options: $TSAny) => {
   const {
@@ -546,11 +552,18 @@ const uploadStudioBackendFiles = async (s3: S3, bucketName: string) => {
       Body: fs.createReadStream(path.join(amplifyDirPath, filePath)),
       Key: path.join(studioBackendDirName, filePath.replace('#current-cloud-backend', '')),
     }));
-  await Promise.all(uploadFileParams.map(params => s3.uploadFile(params)));
+
+  const spinner = ora('Uploading files.');
+  try {
+    spinner.start();
+    await Promise.all(uploadFileParams.map(params => s3.uploadFile(params, false)));
+  } finally {
+    spinner.stop();
+  }
 };
 
 /**
- *
+ * Upload files that Amplify Studio depends on
  */
 export const storeCurrentCloudBackend = async (context: $TSContext) => {
   const zipFilename = '#current-cloud-backend.zip';
@@ -561,6 +574,7 @@ export const storeCurrentCloudBackend = async (context: $TSContext) => {
   if (!fs.existsSync(tempDir)) {
     fs.mkdirSync(tempDir);
   }
+
   // handle tag file
   const tagFilePath = pathManager.getTagFilePath();
   const tagCloudFilePath = pathManager.getCurrentTagFilePath();
@@ -574,11 +588,8 @@ export const storeCurrentCloudBackend = async (context: $TSContext) => {
   });
 
   const zipFilePath = path.normalize(path.join(tempDir, zipFilename));
-  let log = null;
   const result = await archiver.run(currentCloudBackendDir, zipFilePath, undefined, cliJSONFiles);
-
   const s3Key = `${result.zipFilename}`;
-
   const s3 = await S3.getInstance(context);
 
   const s3Params = {
@@ -586,15 +597,9 @@ export const storeCurrentCloudBackend = async (context: $TSContext) => {
     Key: s3Key,
   };
 
-  log = logger('storeCurrentcoudBackend.s3.uploadFile', [{ Key: s3Key }]);
-  log();
-  try {
-    const deploymentBucketName = await s3.uploadFile(s3Params);
-    await uploadStudioBackendFiles(s3, deploymentBucketName);
-  } catch (error) {
-    log(error);
-    throw error;
-  }
+  logger('storeCurrentCloudBackend.s3.uploadFile', [{ Key: s3Key }])();
+  const deploymentBucketName = await s3.uploadFile(s3Params);
+  await uploadStudioBackendFiles(s3, deploymentBucketName);
 
   fs.removeSync(tempDir);
 };
@@ -632,15 +637,8 @@ const prepareResource = async (context: $TSContext, resource: $TSAny) => {
     Body: fs.createReadStream(result.zipFilePath),
     Key: s3Key,
   };
-  const log = logger('packageResources.s3.uploadFile', [{ Key: s3Key }]);
-  log();
-  let s3Bucket: string;
-  try {
-    s3Bucket = await s3.uploadFile(s3Params);
-  } catch (error) {
-    log(error);
-    throw error;
-  }
+  logger('packageResources.s3.uploadFile', [{ Key: s3Key }])();
+  const s3Bucket = await s3.uploadFile(s3Params);
 
   // Update cfn template
   const { category, resourceName }: { category: string; resourceName: string } = resource;
@@ -653,13 +651,10 @@ const prepareResource = async (context: $TSContext, resource: $TSAny) => {
   });
 
   if (cfnFiles.length !== 1) {
-    const errorMessage = cfnFiles.length > 1
-      ? 'Only one CloudFormation template is allowed in the resource directory'
-      : 'CloudFormation template is missing in the resource directory';
-    printer.error(errorMessage);
-    printer.error(resourceDir);
-
-    throw new Error(errorMessage);
+    throw amplifyErrorWithTroubleshootingLink('CloudFormationTemplateError', {
+      message: cfnFiles.length > 1 ? 'Only one CloudFormation template is allowed in the resource directory' : 'No CloudFormation template found in the resource directory',
+      details: `Resource directory: ${resourceDir}`,
+    });
   }
 
   const cfnFile = cfnFiles[0];
@@ -709,23 +704,13 @@ const updateCloudFormationNestedStack = async (
   resourcesToBeUpdated: $TSAny,
   eventMap: $TSAny,
 ) => {
-  const projectRoot = pathManager.findProjectRoot();
-  const backEndDir = pathManager.getBackendDirPath(projectRoot);
-  const rootStackFilePath = path.join(pathManager.getRootStackBuildDirPath(projectRoot), rootStackFileName);
+  const rootStackFilePath = path.join(pathManager.getRootStackBuildDirPath(pathManager.findProjectRoot()), rootStackFileName);
   // deploy preprocess nested stack to disk
   await storeRootStackTemplate(context, nestedStack);
   const transformedStackPath = await preProcessCFNTemplate(rootStackFilePath);
   const cfnItem = await new Cloudformation(context, generateUserAgentAction(resourcesToBeCreated, resourcesToBeUpdated), {}, eventMap);
-  const providerDirectory = path.normalize(path.join(backEndDir, providerName));
 
-  const log = logger('updateCloudFormationNestedStack', [providerDirectory, transformedStackPath]);
-  try {
-    log();
-    await cfnItem.updateResourceStack(transformedStackPath);
-  } catch (error) {
-    log(error);
-    throw error;
-  }
+  await cfnItem.updateResourceStack(transformedStackPath);
 };
 
 const generateUserAgentAction = (resourcesToBeCreated: $TSAny, resourcesToBeUpdated: $TSAny) => {
@@ -847,14 +832,8 @@ export const uploadTemplateToS3 = async (
     Key: `amplify-cfn-templates/${category}/${cfnFile}`,
   };
 
-  const log = logger('uploadTemplateToS3.s3.uploadFile', [{ Key: s3Params.Key }]);
-  let projectBucket: string;
-  try {
-    projectBucket = await s3.uploadFile(s3Params, false);
-  } catch (error) {
-    log(error);
-    throw error;
-  }
+  logger('uploadTemplateToS3.s3.uploadFile', [{ Key: s3Params.Key }])();
+  const projectBucket = await s3.uploadFile(s3Params, false);
 
   if (amplifyMeta) {
     const templateURL = `https://s3.amazonaws.com/${projectBucket}/amplify-cfn-templates/${category}/${cfnFile}`;
@@ -942,7 +921,7 @@ const handleCfnFiles = (
   const { resourceDir, cfnFiles } = getCfnFiles(category, resource);
   cfnFiles.forEach(file => {
     const categoryResources = getCategoryResources(file, resourceDir);
-    // Maping Resource events to categories.
+    // Mapping Resource events to categories.
     categoryResources.forEach(res => {
       eventMap.eventToCategories.set(res, `${category}-${resource}`);
     });
@@ -968,7 +947,7 @@ export const formNestedStack = async (
   // CFN transform for Root stack
   rootStack = await transformRootStack(context);
 
-  // get the {deploymentBucketName , AuthRoleName , UnAuthRole from overridded data}
+  // get the {deploymentBucketName , AuthRoleName , UnAuthRole from overridden data}
 
   const metaToBeUpdated = {
     DeploymentBucketName: rootStack.Resources.DeploymentBucket.Properties.BucketName,
@@ -986,7 +965,7 @@ export const formNestedStack = async (
   const amplifyMeta = useExistingMeta ? projectDetails.amplifyMeta : stateManager.getMeta(projectPath);
   // update amplify meta with updated root stack Info
   if (Object.keys(metaToBeUpdated).length) {
-    context.amplify.updateProvideramplifyMeta(providerName, metaToBeUpdated);
+    context.amplify.updateProviderAmplifyMeta(providerName, metaToBeUpdated);
     // update teamProviderInfo
     const { envName } = context.amplify.getEnvInfo();
     const teamProviderInfo = stateManager.getTeamProviderInfo(projectPath);
@@ -1135,19 +1114,18 @@ export const formNestedStack = async (
               const dependentResource = _.get(amplifyMeta, [dependsOn[i].category, dependsOn[i].resourceName], undefined);
 
               if (!dependentResource && dependsOn[i].category) {
-                throw new Error(`Cannot get resource: ${dependsOn[i].resourceName} from '${dependsOn[i].category}' category.`);
+                throw amplifyErrorWithTroubleshootingLink('PushResourcesError', {
+                  message: `Cannot get resource: ${dependsOn[i].resourceName} from '${dependsOn[i].category}' category.`,
+                });
               }
 
               if (dependentResource && dependentResource.serviceType === 'imported') {
                 const outputAttributeValue = _.get(dependentResource, ['output', attribute], undefined);
 
                 if (!outputAttributeValue) {
-                  const error = new Error(
-                    `Cannot read the '${attribute}' dependent attribute value from the output section of resource: '${dependsOn[i].resourceName}'.`,
-                  );
-                  error.stack = undefined;
-
-                  throw error;
+                  throw amplifyErrorWithTroubleshootingLink('PushResourcesError', {
+                    message: `Cannot read the '${attribute}' dependent attribute value from the output section of resource: '${dependsOn[i].resourceName}'.`,
+                  });
                 }
 
                 parameterValue = outputAttributeValue;